02-mlops-pipeline/requirements-prod.txt
# Core ML Libraries
torch>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Hugging Face & Transformers
transformers>=4.36.0
datasets>=2.15.0
tokenizers>=0.15.0
huggingface-hub>=0.19.0

# MLOps & Orchestration
dagster>=1.5.0
dagster-webserver>=1.5.0

# API and Serving
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pyyaml>=6.0.0
pyarrow>=14.0.0
